requirements.txt
Scrapy==2.11.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
requests==2.31.0
pydantic<2.0.0,>=1.9.1